refactoring, cleanup
