+++ fiftyone/factory/repos/__init__.py
@@ -5,3 +5,7 @@ FiftyOne Server
 | `voxel51.com <https://voxel51.com/>`_
 |
 """
+
+from fiftyone.factory.repos.delegated_operation_doc import (
+    DelegatedOperationDocument,
+)
